backend/requirements.txt
# ── Web Framework ──────────────────────────────────────────────
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1

# ── PDF Parsing ────────────────────────────────────────────────
pdfplumber==0.10.4
python-docx==1.1.0

# ── Database ───────────────────────────────────────────────────
pymongo==4.6.1
motor==3.3.2

# ── NLP & spaCy ───────────────────────────────────────────────
spacy==3.7.4

# ── ML & BERT (Python 3.9 compatible versions) ────────────────
torch==2.2.1
transformers==4.38.2
sentence-transformers==2.5.1
scikit-learn==1.4.1.post1
numpy==1.26.4
pandas==2.2.1

# ── Text Processing ────────────────────────────────────────────
nltk==3.8.1
regex==2023.12.25

# ── Utilities ──────────────────────────────────────────────────
pydantic==2.6.4
